Rename is_degenerate headers to degenerate_shapes
These headers will soon contain other functions related to shape degeneracy.
