+++ include/rlm/cellular/detail/degenerate_shapes.inl
