+++ include/rlm/linear/detail/degenerate_shapes.inl
