+++ test/src/CMakeLists.txt
@@ -34,11 +34,11 @@ target_sources(RlmTest
         "cellular_circle2_tests.cpp"
         "cellular_crop_tests.cpp"
         "cellular_cross_tests.cpp"
+        "cellular_degenerate_shapes_tests.cpp"
         "cellular_distance_between_tests.cpp"
         "cellular_dot_tests.cpp"
         "cellular_shape_edges_tests.cpp"
         "cellular_does_contain_tests.cpp"
-        "cellular_is_degenerate_tests.cpp"
         "cellular_does_fit_tests.cpp"
         "cellular_lerp_tests.cpp"
         "cellular_magnitude_tests.cpp"
@@ -54,7 +54,7 @@ target_sources(RlmTest
         "clamp_tests.cpp"
         "gcf_tests.cpp"
         "lerp_tests.cpp"
-        "linear_is_degenerate_tests.cpp"
+        "linear_degenerate_shapes_tests.cpp"
         "linear_rectangle2_tests.cpp"
         "linear_vector2_tests.cpp"
         "linear_vector3_tests.cpp"
